requirements.txt
boto3>=1.28
botocore>=1.31
watchdog>=3.0
pathspec>=0.11
PyYAML>=6.0
